vm_core.h (iseq_unique_id): prefer uintptr_t instead of unsigned long
It produced a warning about type cast in LLP64 (i.e., windows).

+++ iseq.c
@@ -427,7 +427,7 @@ rb_iseq_memsize(const rb_iseq_t *iseq)
     return size;
 }
 
-static unsigned long fresh_iseq_unique_id = 0; /* -- Remove In 3.0 -- */
+static uintptr_t fresh_iseq_unique_id = 0; /* -- Remove In 3.0 -- */
 
 struct rb_iseq_constant_body *
 rb_iseq_constant_body_alloc(void)
